% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_presence.R
\name{activity_presence}
\alias{activity_presence}
\alias{activity_presence.eventlog}
\alias{activity_presence.grouped_eventlog}
\title{Metric: Activity Presence}
\usage{
activity_presence(eventlog, append)

\method{activity_presence}{eventlog}(eventlog, append = F)

\method{activity_presence}{grouped_eventlog}(eventlog, append = F)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class}

\item{append}{Logical indicating whether to append the result to the original event log
\code{eventlog}.}
}
\description{
Calculates for each activity type in what percentage of cases it is present.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Compute activity presence for event log

\item \code{grouped_eventlog}: Compute activity presence for grouped eventlog
}}

\examples{
\dontrun{
data <- data.frame(case = rep("A",5),
activity_id = c("A","B","C","D","E"),
activity_instance_id = 1:5,
lifecycle_id = rep("complete",5),
timestamp = 1:5,
resource = rep("resource 1", 5))

log <- bupaR::eventlog(data,case_id = "case",
activity_id = "activity_id",
activity_instance_id = "activity_instance_id",
lifecycle_id = "lifecycle_id",
timestamp = "timestamp",
resource_id = "resource")

activity_presence(log)
}
}
