% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/activity_presence_in_cases.R
\name{activity_presence_in_cases}
\alias{activity_presence_in_cases}
\title{Metric: Activity Presence in cases}
\usage{
activity_presence_in_cases(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Calculates for each activity type in what percentage of cases it is present.
}
\examples{
data(example_log)
activity_presence_in_cases(example_log)
}
\seealso{
\code{\link{activity_type_frequency}}
}

