% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/activity_presence_in_traces.R
\name{activity_presence_in_traces}
\alias{activity_presence_in_traces}
\title{Metric: Activity Presence in traces}
\usage{
activity_presence_in_traces(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Calculates for each activity type in what percentage of traces it is present.
}
\examples{
data(example_log)
activity_presence_in_traces(example_log)
}
\seealso{
\code{\link{activity_type_frequency}}
}

