% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/activity_type_frequence.R
\name{activity_type_frequency}
\alias{activity_type_frequency}
\title{Metric: Activity Type Frequency}
\usage{
activity_type_frequency(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of activity type frequency should be performed: trace, case or activity.}
}
\description{
Provides summary statistics about the frequency of activity types at the level of traces, cases or activity types
'
}

