% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_schedule.R
\name{add_fixed_holiday}
\alias{add_fixed_holiday}
\title{Add fixed holiday to work schedule}
\usage{
add_fixed_holiday(work_schedule, name, month, day)
}
\arguments{
\item{work_schedule}{Work schedule created with create_work_schedule}

\item{name}{Name of holiday}

\item{month}{Month in which fixed holiday takes place}

\item{day}{Day of fixed holiday}
}
\description{
Add fixed holiday to work schedule
}
