% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_schedule.R
\name{add_floating_holiday}
\alias{add_floating_holiday}
\title{Add floating holiday to work schedule}
\usage{
add_floating_holiday(work_schedule, name, dates)
}
\arguments{
\item{work_schedule}{Work schedule created with create_work_schedule}

\item{name}{Name of holiday}

\item{dates}{Dates of floating holiday. Make sure to list all dates relevant to your time frame}
}
\description{
Add floating holiday to work schedule
}
