% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_schedule.R
\name{add_holiday_periods}
\alias{add_holiday_periods}
\title{Add holiday period to work schedule}
\usage{
add_holiday_periods(work_schedule, from, to)
}
\arguments{
\item{work_schedule}{Work schedule created with create_work_schedule}

\item{from}{Start of holiday period (included)}

\item{to}{End of holiday period (included)}
}
\description{
Add holiday period to work schedule
}
