% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{BPIC14_incident_case_attributes}
\alias{BPIC14_incident_case_attributes}
\title{BPI Challenge 2014 Incidents}
\format{tbl_df containing 4000 observations with 28 variables}
\usage{
BPIC14_incident_case_attributes
}
\description{
BPI Challenge 2014 Incidents
}
\references{
\url{https://data.3tu.nl/repository/uuid:c3e5d162-0cfd-4bb0-bd82-af5268819c35}
}
\keyword{datasets}

