% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{BPIC15_1}
\alias{BPIC15_1}
\title{BPI Challenge 2015 Log Municipality 1}
\format{eventlog containing 52217 events}
\usage{
BPIC15_1
}
\description{
BPI Challenge 2015 Log Municipality 1
}
\keyword{datasets}

