% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{BPIC15_1_imported}
\alias{BPIC15_1_imported}
\title{BPI Challenge 2015 Log Municipality 1}
\format{eventlog containing 52217 events, as imported from xes}
\usage{
BPIC15_1_imported
}
\description{
BPI Challenge 2015 Log Municipality 1
}
\references{
\url{https://data.3tu.nl/repository/uuid:31a308ef-c844-48da-948c-305d167a0ec1}
}
\keyword{datasets}

