% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_queuing_length.R
\name{calculate_queuing_length}
\alias{calculate_queuing_length}
\title{Calculate queue length}
\usage{
calculate_queuing_length(queueing_times, level, time_interval)
}
\arguments{
\item{queueing_times}{A data frame with queueing times, returned by calculating_queuing_times}

\item{level}{The level of granularity (log, activity or resource). Log: total number of activity instances  that are queued at a given moment in time. Resource: total number activity instances  that are queued for a given resource. Activity: total number of activity instances that are queue for a given activity type.}

\item{time_interval}{The time interval after which the queue length should be calculated. Can be day, week, month, year, or a number of days. The first day for which queue length is calculated is the first timestamp found in the log.}
}
\description{
Calculate queue length
}
