% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case_attributes_from_xes.R
\name{case_attributes_from_xes}
\alias{case_attributes_from_xes}
\title{Case Attributes from Xes-file}
\usage{
case_attributes_from_xes(xesfile = file.choose())
}
\description{
Case Attributes from Xes-file
}

