% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_schedule.R
\name{change_day}
\alias{change_day}
\title{Adjust days in work schedule}
\usage{
change_day(work_schedule, day, start_time, end_time)
}
\arguments{
\item{work_schedule}{Work schedule created with create_work_schedule}

\item{day}{A numeric vector containing the days to be changed. 1 = monday.}

\item{start_time}{The new start time for selected days (hh:mm:ss)}

\item{end_time}{The new end time for selected days (hh:mm:ss)}
}
\description{
Adjust days in work schedule
}
