% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_schedule.R
\name{create_work_schedule}
\alias{create_work_schedule}
\title{Create work schedule}
\usage{
create_work_schedule(start_time = "9:00:00", end_time = "17:00:00")
}
\arguments{
\item{start_time}{Character indicating the usual start time for workdays (hh:mm:ss)}

\item{end_time}{Character indicating the usual end time for workdays (hh:mm:ss)}
}
\description{
Create work schedule
}
