% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{csv_example}
\alias{csv_example}
\title{Example eventlog in data.frame format}
\format{data.frame containing 12766 observations with 5 variables, containg event data}
\usage{
csv_example
}
\description{
Example eventlog in data.frame format
}
\keyword{datasets}

