% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/end_activities.R
\name{end_activities}
\alias{end_activities}
\title{Metric: End activities}
\usage{
end_activities(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of end activities should be performed: log, case or activity.}
}
\description{
At log level, computes how many activity types occur as the last event in a case, both absolute and relative.
At activity level, shows the activities which occur as last, and how often.
The last event in a case is the one which completed the last.
}

