% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventlog_from_xes2.R
\name{eventlog_from_xes2}
\alias{eventlog_from_xes2}
\title{eventlog_from_xes}
\usage{
eventlog_from_xes2(xesfile = file.choose())
}
\arguments{
\item{xesfile}{Reference to a .xes file, conforming to the xes-standard.}
}
\description{
Extracts eventlog from a xes-file.
}
\seealso{
\url{http://www.xes-standard.org/}
}

