% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{example_log}
\alias{example_log}
\title{Example eventlog}
\format{Eventlog containing 12766 events}
\usage{
example_log
}
\description{
Example eventlog
}
\keyword{datasets}

