% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_activity.R
\name{filter_activity}
\alias{filter_activity}
\alias{filter_activity.eventlog}
\alias{filter_activity.grouped_eventlog}
\alias{ifilter_activity}
\title{Filter: Activity}
\usage{
filter_activity(eventlog, activities, reverse)

\method{filter_activity}{eventlog}(eventlog, activities, reverse = FALSE)

\method{filter_activity}{grouped_eventlog}(eventlog, activities,
  reverse = FALSE)

ifilter_activity(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{activities}{A vector of activities to withhold}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters the log based on activities
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter eventlog for activity labels

\item \code{grouped_eventlog}: Filter grouped eventlog for activity labels
}}

