% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_activity_instances.R
\name{filter_activity_instance}
\alias{filter_activity_instance}
\alias{ifilter_activity_instance}
\title{title Filter: Activity instance}
\usage{
filter_activity_instance(eventlog, activity_instances = NULL, reverse = FALSE)

ifilter_activity_instance(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{activity_instances}{A vector of activity instance identifiers}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Filters the log based on activity instance identifier
}
\details{
The method filter_activity_instance can be used to filter on activity instance identifiers. It has an activity_instances argument,
to which a vector of identifiers can be given. The selection can be negated with the reverse argument.
}
\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{vignette("filters", "edeaR")}

Other filters: 
\code{\link{filter_activity_frequency}()}
}
