% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_activity_presence.R
\name{filter_activity_presence}
\alias{filter_activity_presence}
\alias{filter_activity_presence.eventlog}
\alias{filter_activity_presence.grouped_eventlog}
\alias{ifilter_activity_presence}
\title{Filter: Activity Presence}
\usage{
filter_activity_presence(eventlog, activities, method, reverse)

\method{filter_activity_presence}{eventlog}(eventlog, activities = NULL,
  method = c("all", "one_of", "none"), reverse = FALSE)

\method{filter_activity_presence}{grouped_eventlog}(eventlog,
  activities = NULL, method = c("all", "one_of", "none"), reverse = FALSE)

ifilter_activity_presence(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{activities}{A vector of activities to withhold.}

\item{method}{Filter method. If "all", each of the activities should be present. If "one_of", at least one of them should be present. If "none", none of the activities are allowed to occur in the filtered traces.}
}
\description{
Filters cases based on the presence (or absence) of activities
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log on presence of activities.

\item \code{grouped_eventlog}: Filter grouped event log on presence of activities.
}}

