% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_attributes.R
\name{filter_attributes}
\alias{filter_attributes}
\title{Generic filter function for eventlog}
\usage{
filter_attributes(eventlog, ...)
}
\arguments{
\item{eventlog}{Eventlog object}

\item{...}{Filter conditions}
}
\description{
Generic filter function for eventlog
}
