% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_case.R
\name{filter_case}
\alias{filter_case}
\alias{filter_case.eventlog}
\alias{filter_case.grouped_eventlog}
\alias{ifilter_case}
\title{title Filter: Case}
\usage{
filter_case(eventlog, cases = NULL, reverse = FALSE)

\method{filter_case}{eventlog}(eventlog, cases = NULL, reverse = FALSE)

\method{filter_case}{grouped_eventlog}(eventlog, cases = NULL, reverse = FALSE)

ifilter_case(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{cases}{A vector of cases identifiers}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Filters the log based on case identifier
}
\details{
The method filter_case can be used to filter on case identifiers. It has an cases argument,
to which a vector of identifiers can be given. The selection can be negated with the reverse argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter for eventlogs

\item \code{grouped_eventlog}: Stratified filter for grouped eventlogs
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{vignette("filters", "edeaR")}

Other filters: 
\code{\link{filter_activity_frequency}()}
}
