% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_case_condition.R
\name{filter_case_condition}
\alias{filter_case_condition}
\title{title Filter: Case}
\usage{
filter_case_condition(eventlog, condition = NULL, reverse = FALSE)
}
\arguments{
\item{eventlog}{The dataset to be used. Should be a (grouped) eventlog object.}

\item{condition}{A condition}

\item{reverse}{Logical, indicating whether the selection should be reversed.}
}
\value{
When given an eventlog, it will return a filtered eventlog. When given a grouped eventlog, the filter will be applied
in a stratified way (i.e. each separately for each group). The returned eventlog will be grouped on the same variables as
the original event log.
}
\description{
Filters cases using a condition
}
\details{
Only keeps cases if the condition is valid for at least one event.
}
\seealso{
\code{vignette("filters", "edeaR")}
}
