% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_resource.R
\name{filter_resource}
\alias{filter_resource}
\alias{filter_resource.eventlog}
\alias{filter_resource.grouped_eventlog}
\alias{ifilter_resource}
\title{Filter: Resource}
\usage{
filter_resource(eventlog, resources, reverse)

\method{filter_resource}{eventlog}(eventlog, resources, reverse = FALSE)

\method{filter_resource}{grouped_eventlog}(eventlog, resources,
  reverse = FALSE)

ifilter_resource(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{resources}{A vector of resources to withhold}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters the log based on resources
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log

\item \code{grouped_eventlog}: Filter grouped event log
}}

