% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_resource_frequency.R
\name{filter_resource_frequency}
\alias{filter_resource_frequency}
\alias{ifilter_resource_frequency}
\title{Filter: Resource frequency}
\usage{
filter_resource_frequency(eventlog, percentile_cut_off = 0.8, reverse = F)

ifilter_resource_frequency(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{percentile_cut_off}{The target coverage of events
A percentile of 0.9 will return the most common resource types of the eventlog, which account for 90\% of the events.}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters the log based on its most frequent resources, until a specific percentile cut off.
}
