% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_trace.R
\name{filter_trace}
\alias{filter_trace}
\title{title Filter: Trace}
\usage{
filter_trace(eventlog, trace_ids, reverse)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{trace_ids}{A vector of trace identifiers}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Filters the log based on trace id
}
\details{
The method filter_trace can be used to filter on trace id It has an trace_ids argument,
to which a vector of identifiers can be given. The selection can be negated with the reverse argument.
}
\seealso{
\code{vignette("filters", "edeaR")}
}
