% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_trim.R
\name{filter_trim}
\alias{filter_trim}
\alias{filter_trim.eventlog}
\alias{filter_trim.grouped_eventlog}
\alias{ifilter_trim}
\title{Filter: Trim cases}
\usage{
filter_trim(eventlog, start_activities, end_activities, reverse)

\method{filter_trim}{eventlog}(eventlog, start_activities = NULL,
  end_activities = NULL, reverse = FALSE)

\method{filter_trim}{grouped_eventlog}(eventlog, start_activities = NULL,
  end_activities = NULL, reverse = FALSE)

ifilter_trim(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{start_activities}{Start activities used for trimming. If not provided, the start of the cases is not trimmed.}

\item{end_activities}{End activities used for trimming. If not provided, the end of the cases or not trimmed.}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Trim all cases from the first event of a set of start activities to the last event of a set of end activities.
Traces that don't have at least one event of both sets are discarded.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log

\item \code{grouped_eventlog}: Filter grouped event log
}}

