% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_activity.R,
%   R/filter_activity_frequency.R, R/filter_activity_instance.R,
%   R/filter_activity_presence.R, R/filter_case.R, R/filter_endpoints.R,
%   R/filter_lifecycle.R, R/filter_lifecycle_presence.R,
%   R/filter_precedence.R, R/filter_processing_time.R, R/filter_resource.R,
%   R/filter_resource_frequency.R, R/filter_throughput_time.R,
%   R/filter_time_period.R, R/filter_trace_frequency.R, R/filter_trace_length.R,
%   R/filter_trim.R, R/filter_trim_lifecycle.R, R/ifilter.R
\name{ifilter_activity}
\alias{ifilter_activity}
\alias{ifilter_activity_frequency}
\alias{ifilter_activity_instance}
\alias{ifilter_activity_presence}
\alias{ifilter_case}
\alias{ifilter_endpoints}
\alias{ifilter_lifecycle}
\alias{ifilter_lifecycle_presence}
\alias{ifilter_precedence}
\alias{ifilter_processing_time}
\alias{ifilter_resource}
\alias{ifilter_resource_frequency}
\alias{ifilter_throughput_time}
\alias{ifilter_time_period}
\alias{ifilter_trace_frequency}
\alias{ifilter_trace_length}
\alias{ifilter_trim}
\alias{ifilter_trim_lifecycle}
\alias{ifilter}
\title{ifilter functions}
\usage{
ifilter_activity(eventlog)

ifilter_activity_frequency(eventlog)

ifilter_activity_instance(eventlog)

ifilter_activity_presence(eventlog)

ifilter_case(eventlog)

ifilter_endpoints(eventlog)

ifilter_lifecycle(eventlog)

ifilter_lifecycle_presence(eventlog)

ifilter_precedence(eventlog)

ifilter_processing_time(eventlog)

ifilter_resource(eventlog)

ifilter_resource_frequency(eventlog)

ifilter_throughput_time(eventlog)

ifilter_time_period(eventlog)

ifilter_trace_frequency(eventlog)

ifilter_trace_length(eventlog)

ifilter_trim(eventlog)

ifilter_trim_lifecycle(eventlog)
}
\description{
ifilter functions
}
\keyword{internal}
\keyword{internals}
