% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/life_cycle_id.R
\name{life_cycle_id}
\alias{life_cycle_id}
\title{Life cycle classifier}
\usage{
life_cycle_id(eventlog)
}
\arguments{
\item{eventlog}{An object of class \code{eventlog}.}
}
\description{
Get the life_cycle_id of an object of class \code{eventlog}
}
\details{

}
\seealso{
\code{\link{eventlog}}, \code{\link{activity_instance_id}}
}

