% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_repetitions.R
\name{number_of_repetitions}
\alias{number_of_repetitions}
\title{Metric:  Number of repetitions}
\usage{
number_of_repetitions(eventlog, type = c("repeat", "redo"),
  level_of_analysis = c("log", "case", "activity", "resource",
  "resource-activity"))
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{type}{The type of repetitions, either repeat or redo}

\item{level_of_analysis}{At which level the analysis of repetitions should be performed: log, case, activity, resource, resource-activity.}
}
\description{
Provides summuary statistics on the number of repetitions, at the level of activity types, traces, cases and the eventlog.
}
