% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/number_of_traces_with_selfloop.R
\name{number_of_traces_with_selfloop}
\alias{number_of_traces_with_selfloop}
\title{Metric: Number of Traces with Selfloop}
\usage{
number_of_traces_with_selfloop(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Returns the number of traces in which one or more selfloops occur, both in absolute and relative numbers.
}

