% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.activity_frequency.R
\name{plot.activity_frequency}
\alias{plot.activity_frequency}
\title{Activity Frequency Plot}
\usage{
\method{plot}{activity_frequency}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize activity freqeuncy data.
}
