% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.activity_presence.R
\name{plot.activity_presence}
\alias{plot.activity_presence}
\title{Activity Presence Plot}
\usage{
\method{plot}{activity_presence}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize activity presence data.
}
