% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.end_activity.R
\name{plot.end_activities}
\alias{plot.end_activities}
\title{End Activities Plot}
\usage{
\method{plot}{end_activities}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize end activities data.
}
