% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.idle_time.R
\name{plot.idle_time}
\alias{plot.idle_time}
\title{Idle Time Plot}
\usage{
\method{plot}{idle_time}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize idle time data.
}
