% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.processing_time.R
\name{plot.processing_time}
\alias{plot.processing_time}
\title{PRocessing time Plot}
\usage{
\method{plot}{processing_time}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize process time data.
}
