% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.referral_matrix.R
\name{plot.referral_matrix}
\alias{plot.referral_matrix}
\title{Referral Matrix Plot}
\usage{
\method{plot}{referral_matrix}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize referral matrix
}
