% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resource_frequency.R
\name{plot.resource_frequency}
\alias{plot.resource_frequency}
\title{Resource frequency Plot}
\usage{
\method{plot}{resource_frequency}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize resource frequency data.
}
