% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resource_involvement.R
\name{plot.resource_involvement}
\alias{plot.resource_involvement}
\title{Resource Involvement Plot}
\usage{
\method{plot}{resource_involvement}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize resource involvement data.
}
