% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.resource_specialisation.R
\name{plot.resource_specialisation}
\alias{plot.resource_specialisation}
\title{Resource Specialisation Plot}
\usage{
\method{plot}{resource_specialisation}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize resource specialisation data.
}
