% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.start_activities.R
\name{plot.start_activities}
\alias{plot.start_activities}
\title{Start Activities Plot}
\usage{
\method{plot}{start_activities}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize start activities data.
}
