% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.throughput_time.R
\name{plot.throughput_time}
\alias{plot.throughput_time}
\title{Throughput time Plot}
\usage{
\method{plot}{throughput_time}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize throughput time data.
}
