% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trace_coverage.R
\name{plot.trace_coverage}
\alias{plot.trace_coverage}
\title{Trace Coverage Plot}
\usage{
\method{plot}{trace_coverage}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize trace coverage data.
}
