% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.trace_length.R
\name{plot.trace_length}
\alias{plot.trace_length}
\title{Trace Length Plot}
\usage{
\method{plot}{trace_length}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize trace length data.
}
