% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.activity_frequency.R
\name{print.activity_frequency}
\alias{print.activity_frequency}
\title{Activity Frequency Print}
\usage{
\method{print}{activity_frequency}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Actvitity Frequency Information
}
