% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.idle_time.R
\name{print.idle_time}
\alias{print.idle_time}
\title{Idle Time Print}
\usage{
\method{print}{idle_time}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print idle time Information
}
