% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.number_of_repetitions.R
\name{print.number_of_repetitions}
\alias{print.number_of_repetitions}
\title{Repetitions Print}
\usage{
\method{print}{number_of_repetitions}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Repetitions Information
}
