% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.number_of_selfloops.R
\name{print.number_of_selfloops}
\alias{print.number_of_selfloops}
\title{Selfloops Print}
\usage{
\method{print}{number_of_selfloops}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Selfloops Information
}
