% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.processing_time.R
\name{print.processing_time}
\alias{print.processing_time}
\title{Processing Time Print}
\usage{
\method{print}{processing_time}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Processing Time Information
}
