% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.throughput_time.R
\name{print.throughput_time}
\alias{print.throughput_time}
\title{Throughput Time Print}
\usage{
\method{print}{throughput_time}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Throughput Time Information
}
