% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trace_coverage.R
\name{print.trace_coverage}
\alias{print.trace_coverage}
\title{Trace coverage print}
\usage{
\method{print}{trace_coverage}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Trace coverage Information
}
