% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.trace_length.R
\name{print.trace_length}
\alias{print.trace_length}
\title{Trace length Print}
\usage{
\method{print}{trace_length}(x, ...)
}
\arguments{
\item{x}{Data to print}

\item{...}{Additional arguments}
}
\description{
Print Trace length Information
}
