% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_schedule.R
\name{print.work_schedule}
\alias{print.work_schedule}
\title{Print work schedule}
\usage{
\method{print}{work_schedule}(x, ...)
}
\arguments{
\item{x}{Work schedule to print}

\item{...}{Additional arguments (ignored)}
}
\description{
Print work schedule
}
