% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repetitions.R
\name{repetitions}
\alias{repetitions}
\title{Metric:  Repetitions}
\usage{
repetitions(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of repetitions should be performed: log, case, trace or activity.}
}
\description{
Provides summuary statistics on the number of repetitions, at the level of activity types, traces, cases and the eventlog.
}

