% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_activity_involvement.R
\name{resource_activity_involvement}
\alias{resource_activity_involvement}
\title{Metric: Resource Activity Involvement}
\usage{
resource_activity_involvement(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Calculates for each resource in what percentage of cases it perform a specific activity.
}

