% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_frequency.R
\name{resource_frequency}
\alias{resource_frequency}
\title{Metric: Resource frequency}
\usage{
resource_frequency(eventlog, level_of_analysis, include_zeros = F)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of  coverage should be performed: log, case, activity, resource, resource-activity.}

\item{include_zeros}{At the resource-activity level, included pairs which do not occur.}
}
\description{
Analyses the frequency of resources at different levels of analysis
}

