% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_involvement.R
\name{resource_involvement}
\alias{resource_involvement}
\title{Metric: Resource Involvement}
\usage{
resource_involvement(eventlog, level_of_analysis = c("case", "resource",
  "resource-activity"))
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{The level of analysis: resource or resource-activity.}
}
\description{
Calculates for each resource/resource-activity in what percentage of cases it is present.
}
