% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_specialization.R
\name{resource_specialization}
\alias{resource_specialization}
\title{Metric: Resource Specialization}
\usage{
resource_specialization(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of  coverage should be performed: log, case, resource.}
}
\description{
Analyses whether resources specialize in specific activities
}

