% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\name{setdiff}
\alias{setdiff}
\alias{setdiff.eventlog}
\title{Set operations}
\usage{
setdiff(x, y, ...)

\method{setdiff}{eventlog}(x, y, ...)
}
\arguments{
\item{x}{Eventlog}

\item{y}{Eventlog}
}
\value{
Eventlog of activity instances in x which do not belong to y
}
\description{
Set operations
}
\section{Methods (by class)}{
\itemize{
\item \code{setdiff(eventlog)}: Setdiff for eventlogs

}}
