% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.eventlog.R
\name{summary.eventlog}
\alias{summary.eventlog}
\title{Generic summary function for eventlog class}
\usage{
\method{summary}{eventlog}(object, ...)
}
\arguments{
\item{...}{Additional Arguments}
}
\description{
Generic summary function for eventlog class
 @param object Eventlog object
}

