% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace_coverage.R
\name{trace_coverage}
\alias{trace_coverage}
\title{Metric: Trace Coverage}
\usage{
trace_coverage(eventlog, threshold = 0.8)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{threshold}{The percentage of cases to cover.}
}
\description{
Computes how many traces are needed to cover a certain percentage of the log.
When a tie exist, the two nearest breakpoints are returned.
}

