% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trace_frequency.R
\name{trace_frequency}
\alias{trace_frequency}
\title{Metric: Trace Frequency}
\usage{
trace_frequency(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Computes the absolute and relative frequency of the traces in the event log.
#'
}

