% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_xes.R
\name{write_xes}
\alias{write_xes}
\title{Write XES file}
\usage{
write_xes(eventlog, case_attributes = NULL, file = file.choose())
}
\arguments{
\item{eventlog}{An event log object}

\item{case_attributes}{Data object containing case attributes}

\item{file}{Destination file}
}
\description{
Function for writing xes-file
}

