% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_type_frequence.R
\name{activity_frequency}
\alias{activity_frequency}
\title{Metric: Activity Frequency}
\usage{
activity_frequency(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of activity type frequency should be performed: log, trace, case, activity.}
}
\description{
Provides summary statistics about the frequency of activity types at the level of traces, cases, resources or activity types.
'
}

