% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/activity_presence.R
\name{activity_presence}
\alias{activity_presence}
\title{Metric: Activity Presence}
\usage{
activity_presence(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Calculates for each activity type in what percentage of cases it is present.
}
\examples{
data(example_log)
activity_presence(example_log)
}
\seealso{
\code{\link{activity_type_frequency}}
}

