% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_presence.R
\name{activity_presence}
\alias{activity_presence}
\alias{activity_presence.eventlog}
\alias{activity_presence.grouped_eventlog}
\alias{activity_presence.activitylog}
\alias{activity_presence.grouped_activitylog}
\title{Metric: Activity Presence}
\usage{
activity_presence(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  ...,
  eventlog = deprecated()
)

\method{activity_presence}{eventlog}(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  ...,
  eventlog = deprecated()
)

\method{activity_presence}{grouped_eventlog}(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  ...,
  eventlog = deprecated()
)

\method{activity_presence}{activitylog}(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  ...,
  eventlog = deprecated()
)

\method{activity_presence}{grouped_activitylog}(
  log,
  append = deprecated(),
  append_column = NULL,
  sort = TRUE,
  ...,
  eventlog = deprecated()
)
}
\arguments{
\item{log}{\code{\link[bupaR]{log}}: Object of class \code{\link[bupaR]{log}}, \code{\link[bupaR]{grouped_log}}, \code{\link[bupaR]{eventlog}}, \code{\link[bupaR]{activitylog}}, etc.}

\item{append}{\code{\link{logical}} (default \code{FALSE}) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}: Indicating whether to append results to original log. Ignored when level is \code{log} or \code{trace}.}

\item{append_column}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Which of the output columns to append to log, if \code{append = TRUE}. Default column depends on chosen level.}

\item{sort}{\code{\link{logical}} (default \code{TRUE}): Sort output on count. Only for levels with frequency count output.}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated arguments}

\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Calculates for each activity type in what percentage of cases it is present.
}
\details{
An indication of variance can be the presence of the activities in the different cases. This metric shows for each activity the absolute
number of cases in which each activity occurs together with its relative presence.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Compute activity presence for an \code{\link[bupaR]{eventlog}}.

\item \code{grouped_eventlog}: Compute activity presence for a \code{\link[bupaR]{grouped_eventlog}}.

\item \code{activitylog}: Compute activity presence for an \code{\link[bupaR]{activitylog}}.

\item \code{grouped_activitylog}: Compute activity presence for a \code{\link[bupaR]{grouped_activitylog}}.
}}

\examples{
\dontrun{
data <- data.frame(case = rep("A",5),
activity_id = c("A","B","C","D","E"),
activity_instance_id = 1:5,
lifecycle_id = rep("complete",5),
timestamp = 1:5,
resource = rep("resource 1", 5))

log <- bupaR::eventlog(data,case_id = "case",
activity_id = "activity_id",
activity_instance_id = "activity_instance_id",
lifecycle_id = "lifecycle_id",
timestamp = "timestamp",
resource_id = "resource")

activity_presence(log)
}

}
\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
Other metrics: 
\code{\link{activity_frequency}()},
\code{\link{end_activities}()}
}
\concept{metrics}
