% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_specialization.R
\name{activity_specialization}
\alias{activity_specialization}
\title{Metric: Activity Specialization}
\usage{
activity_specialization(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of  coverage should be performed: case, activity}
}
\description{
Analyses whether activities are specialize in by specific resources
}

