% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{BPIC14_incident_case_attributes}
\alias{BPIC14_incident_case_attributes}
\title{BPI Challenge 2014 Incidents}
\format{tbl_df containing 4000 observations with 28 variables}
\usage{
BPIC14_incident_case_attributes
}
\keyword{datasets}

