% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{BPIC14_incident_log}
\alias{BPIC14_incident_log}
\title{BPI Challenge 2014 Incident Activities}
\format{eventlog containing 39911 events}
\usage{
BPIC14_incident_log
}
\keyword{datasets}

