% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_queuing_times.R
\name{calculate_queuing_times}
\alias{calculate_queuing_times}
\title{Calculate queuing times}
\usage{
calculate_queuing_times(eventlog)
}
\arguments{
\item{eventlog}{The event log object}
}
\value{
Returns a list of all the activity instances, with the time they started, and the time since they were queued. Notice that this does not take into account
any process model notion! The time since they are queued is the completion time of the previous activity in the log.
}
\description{
Calculate queuing times
}
