% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_attributes_from_xes.R
\name{case_attributes_from_xes}
\alias{case_attributes_from_xes}
\title{Case Attributes from Xes-file}
\usage{
case_attributes_from_xes(xesfile = file.choose())
}
\arguments{
\item{xesfile}{Reference to a .xes file, conforming to the xes-standard.}
}
\description{
Extracts case attributes from a xes-file.
}
\seealso{
\url{http://www.xes-standard.org/}
}

