% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edeaR.R
\docType{package}
\name{edeaR}
\alias{edeaR-package}
\alias{edeaR}
\title{edeaR - Exploratory and Descriptive Event-based data Analysis in R}
\description{
This package provides several useful techniques for Exploratory and Descriptive analysis of event based data in R, developed by the Business Informatics Research Group of Hasselt University.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bupar.net/}
  \item \url{https://github.com/bupaverse/edeaR/}
  \item \url{https://bupaverse.github.io/edeaR/}
  \item Report bugs at \url{https://github.com/bupaverse/edeaR/issues/}
}

}
\author{
\strong{Maintainer}: Gert Janssenswillen \email{gert.janssenswillen@uhasselt.be}

Other contributors:
\itemize{
  \item Gerard van Hulzen \email{gerard.vanhulzen@uhasselt.be} [contributor]
  \item Marijke Swennen \email{marijke.swennen@uhasselt.be} [contributor]
  \item Ivan Esin \email{ivan.esin@student.uhasselt.be} [contributor]
  \item Hasselt University [copyright holder]
}

}
\keyword{internal}
