% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edeaR.R
\docType{package}
\name{edeaR}
\alias{edeaR}
\alias{edeaR-package}
\title{edeaR - Exploratory and Descriptive Event-based data Analysis in R}
\description{
This package provides several useful techniques for Exploratory and Descriptive analysis of event based data in R, developed by the Business Informatics Research Group of Hasselt University.
}
