% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_activity.R
\name{filter_activity}
\alias{filter_activity}
\alias{ifilter_activity}
\title{Filter: Activity}
\usage{
filter_activity(eventlog, activities = NULL, reverse = F)

ifilter_activity(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{activities}{A vector of activities to withhold}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters the log based on activities
}
