% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_activity_presence.R
\name{filter_activity_presence}
\alias{filter_activity_presence}
\title{Filter: Activity Presence}
\usage{
filter_activity_presence(eventlog, activities = NULL, method = "all")
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{activities}{A vector of activities to withhold.}

\item{method}{Filter method. If "all", each of the activities should be present. If "one_of", at least one of them should be present. If "none",}
}
\description{
Filters cases based on the presence (or absence) of activities
}

