% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_case.R
\name{filter_case}
\alias{filter_case}
\title{Filter: Case}
\usage{
filter_case(eventlog, cases = NULL, reverse = F)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{cases}{A vector of cases to withhold}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters the log based on case identifier
}

