% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_case.R
\name{filter_case}
\alias{filter_case}
\alias{filter_case.eventlog}
\alias{filter_case.grouped_eventlog}
\alias{ifilter_case}
\title{Filter: Case}
\usage{
filter_case(eventlog, cases, reverse)

\method{filter_case}{eventlog}(eventlog, cases = NULL, reverse = F)

\method{filter_case}{grouped_eventlog}(eventlog, cases = NULL, reverse = F)

ifilter_case(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{cases}{A vector of cases to withhold}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters the log based on case identifier
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter for eventlogs

\item \code{grouped_eventlog}: Stratified filter for grouped eventlogs
}}

