% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_endpoints.R
\name{filter_endpoints}
\alias{filter_endpoints}
\alias{filter_endpoints.eventlog}
\alias{filter_endpoints.grouped_eventlog}
\alias{ifilter_endpoints}
\title{Filter: Start and end activities}
\usage{
filter_endpoints(
  eventlog,
  start_activities = NULL,
  end_activities = NULL,
  percentage = NULL,
  reverse = FALSE,
  ...
)

\method{filter_endpoints}{eventlog}(
  eventlog,
  start_activities = NULL,
  end_activities = NULL,
  percentage = NULL,
  reverse = FALSE,
  ...
)

\method{filter_endpoints}{grouped_eventlog}(
  eventlog,
  start_activities = NULL,
  end_activities = NULL,
  percentage = NULL,
  reverse = FALSE,
  ...
)

ifilter_endpoints(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{start_activities}{A vector of activity identifiers, or NULL}

\item{end_activities}{A vector of activity identifiers, or NULL}

\item{percentage}{A percentage p to be used as percentile cut off. When this is used, the most common endpoint-pairs will be selected until at least
the p\\% of the cases are selected.}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated arguments}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Filters the log based on a provided set of start and end activities
}
\details{
The filter_endpoints method filters cases based on the first and last activity label. It can be used in two ways: by specifying vectors with allowed start
activities and/or allowed end activities, or by specifying a percentile. In the latter case, the percentile value will be used as a cut off.
For example, when set to 0.9, it will select the most common endpoint pairs which together cover at least 90\% of the cases, and filter the event log accordingly.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log

\item \code{grouped_eventlog}: Filter grouped event log stratified
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{vignette("filters", "edeaR")}

Other filters: 
\code{\link{filter_activity_frequency}()},
\code{\link{filter_activity_instance}()},
\code{\link{filter_activity_presence}()},
\code{\link{filter_case}()}
}
