% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_infrequent_flows.R
\name{filter_infrequent_flows}
\alias{filter_infrequent_flows}
\title{Filter cases based on infrequent flows.}
\usage{
filter_infrequent_flows(eventlog, min_n)
}
\arguments{
\item{eventlog}{The event log to use}

\item{min_n}{Cases containing a flow that occurs less than min_n times are discarded.}
}
\value{
Filtered event log.
}
\description{
Filter cases based on infrequent flows.
}
