% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_resource.R
\name{filter_resource}
\alias{filter_resource}
\alias{filter_resource.eventlog}
\alias{filter_resource.grouped_eventlog}
\alias{ifilter_resource}
\title{Filter: Resource}
\usage{
filter_resource(eventlog, resources, reverse = FALSE)

\method{filter_resource}{eventlog}(eventlog, resources, reverse = FALSE)

\method{filter_resource}{grouped_eventlog}(eventlog, resources, reverse = FALSE)

ifilter_resource(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{resources}{A vector of resources identifiers}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Filters the log based on resource identifiers
}
\details{
#' The method filter_resource can be used to filter on resource identifiers. It has a resources argument,
to which a vector of identifiers can be given. The selection can be negated with the reverse argument.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log

\item \code{grouped_eventlog}: Filter grouped event log
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{vignette("filters", "edeaR")}

Other filters: 
\code{\link{filter_activity_frequency}()},
\code{\link{filter_activity_instance}()},
\code{\link{filter_activity_presence}()},
\code{\link{filter_case_condition}()},
\code{\link{filter_case}()},
\code{\link{filter_endpoints_conditions}()},
\code{\link{filter_endpoints}()}
}
