% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_trace.R
\name{filter_trace}
\alias{filter_trace}
\title{title Filter: Trace}
\usage{
filter_trace(eventlog, trace_ids, reverse)
}
\arguments{
\item{eventlog}{The dataset to be used. Should be a (grouped) eventlog object.}

\item{trace_ids}{A vector of trace identifiers}

\item{reverse}{Logical, indicating whether the selection should be reversed.}
}
\value{
When given an eventlog, it will return a filtered eventlog. When given a grouped eventlog, the filter will be applied
in a stratified way (i.e. each separately for each group). The returned eventlog will be grouped on the same variables as
the original event log.
}
\description{
Filters the log based on trace id
}
\details{
The method filter_trace can be used to filter on trace id It has an trace_ids argument,
to which a vector of identifiers can be given. The selection can be negated with the reverse argument.
}
\seealso{
\code{vignette("filters", "edeaR")}
}
