% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_trace_length.R
\name{filter_trace_length}
\alias{filter_trace_length}
\title{Filter: Trace length percentile}
\usage{
filter_trace_length(eventlog, lower_threshold = NULL,
  upper_threshold = NULL, percentile_cut_off = NULL, reverse = F)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{lower_threshold}{The lower trace length threshold.
When \code{reverse} is FALSE, all traces with a lower frequency are discarded.}

\item{upper_threshold}{The upper trace length threshold.
When \code{reverse} is FALSE, all traces with a lo frequency are discarded.}

\item{percentile_cut_off}{Alternatively to providing thresholds, a percentile cut off can be provided.
A percentile cut off value of 0.9 will return the 90\% shortest cases.
When \code{reverse} is set to TRUE, it will return the 10\% longest cases.}

\item{reverse}{A logical parameter depicting whether the selection should be reversed.}
}
\description{
Filters cases on length, using a percentile threshold.
}

