% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_trim.R
\name{filter_trim}
\alias{filter_trim}
\alias{filter_trim.eventlog}
\alias{filter_trim.grouped_eventlog}
\alias{ifilter_trim}
\title{Filter: Trim cases}
\usage{
filter_trim(
  eventlog,
  start_activities = NULL,
  end_activities = NULL,
  reverse = FALSE
)

\method{filter_trim}{eventlog}(
  eventlog,
  start_activities = NULL,
  end_activities = NULL,
  reverse = FALSE
)

\method{filter_trim}{grouped_eventlog}(
  eventlog,
  start_activities = NULL,
  end_activities = NULL,
  reverse = FALSE
)

ifilter_trim(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{start_activities, end_activities}{\code{\link{character}} vector (default \code{\link{NULL}}): A vector of activity identifiers, or \code{\link{NULL}}.}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Trim cases from the first event of a set of start activities to the last event of a set of end activities.
}
\details{
One can trim cases by removing one or more activity instances
at the start and/or end of a case. Trimming is performed until all cases have a start and/or
end point belonging to a set of allowed activity labels. This filter requires a set of allowed
start activities and/or a set of allowed end activities. If one of them is not provided it will
not trim the cases at this edge.The selection can be reversed, which means that
only the trimmed events at the start and end of cases are retained. As such, this argument
allows to cut intermediate parts out of traces.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log

\item \code{grouped_eventlog}: Filter grouped event log
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{vignette("filters", "edeaR")}

Other filters: 
\code{\link{filter_activity_frequency}()},
\code{\link{filter_activity_instance}()},
\code{\link{filter_activity_presence}()},
\code{\link{filter_case_condition}()},
\code{\link{filter_case}()},
\code{\link{filter_endpoints_conditions}()},
\code{\link{filter_endpoints}()}
}
