% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_lifecycle_trim.R
\name{filter_trim_lifecycle}
\alias{filter_trim_lifecycle}
\alias{filter_trim_lifecycle.eventlog}
\alias{filter_trim_lifecycle.grouped_eventlog}
\alias{ifilter_trim_lifecycle}
\title{Filter: Trim activity instances based on life cycle labels}
\usage{
filter_trim_lifecycle(
  eventlog,
  start_lifecycle = NULL,
  end_lifecycle = NULL,
  reverse = FALSE
)

\method{filter_trim_lifecycle}{eventlog}(
  eventlog,
  start_lifecycle = NULL,
  end_lifecycle = NULL,
  reverse = FALSE
)

\method{filter_trim_lifecycle}{grouped_eventlog}(
  eventlog,
  start_lifecycle = NULL,
  end_lifecycle = NULL,
  reverse = FALSE
)

ifilter_trim_lifecycle(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{start_lifecycle}{A vector of life cycle identifiers, or NULL}

\item{end_lifecycle}{A vector of life cycle identifiers, or NULL}

\item{reverse}{\code{\link{logical}} (default \code{FALSE}): Indicating whether the selection should be reversed.}
}
\value{
When given an object of type \code{\link[bupaR]{log}}, it will return a filtered \code{\link[bupaR]{log}}.
When given an object of type \code{\link[bupaR]{grouped_log}}, the filter will be applied in a stratified way (i.e. each separately for each group).
The returned log will be grouped on the same variables as the original log.
}
\description{
Trim activity instances from the first event of a set of start life cycle labels to the last event of a set of end life cycle labels
}
\details{
One can trim activity instances  by removing one or more events at the start and/or end of the activity instances.
Trimming is performed until all activity instances have a start and/or
end point belonging to a set of allowed life cycle labels. This filter requires a set of allowed
start life cycle labels and/or a set of allowed life cycle labels. If one of them is not provided it will
not trim the activity instances at this edge.The selection can be reversed, which means that
only the trimmed events at the start and end of activity instances are retained. As such, this argument
allows to cut intermediate parts out of activity instances.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Filter event log

\item \code{grouped_eventlog}: Filter grouped event log
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{vignette("filters", "edeaR")}

Other filters: 
\code{\link{filter_activity_frequency}()},
\code{\link{filter_activity_instance}()},
\code{\link{filter_activity_presence}()},
\code{\link{filter_case_condition}()},
\code{\link{filter_case}()},
\code{\link{filter_endpoints_conditions}()},
\code{\link{filter_endpoints}()},
\code{\link{filter_precedence_condition}()},
\code{\link{filter_precedence_resource}()},
\code{\link{filter_precedence}()}
}
