% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/number_of_selfloops.R
\name{number_of_selfloops}
\alias{number_of_selfloops}
\title{Metric: Number of selfloops in trace}
\usage{
number_of_selfloops(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of selfloops should be performed: trace or log}
}
\description{
Returns the number of selfloops in each trace. Can be performed at the level of traces, activities, or the level of the event log.
}

