% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_traces.R
\name{number_of_traces}
\alias{number_of_traces}
\alias{number_of_traces.eventlog}
\alias{number_of_traces.grouped_eventlog}
\title{Metric: Number of traces}
\usage{
number_of_traces(eventlog)

\method{number_of_traces}{eventlog}(eventlog)

\method{number_of_traces}{grouped_eventlog}(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Computes how many traces there are.
}
\details{
This metric provides two values, the absolute and relative number of traces that
occur in the event log. The relative number shows expected number of traces needed to cover 100 cases.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Number of traces in eventlog

\item \code{grouped_eventlog}: Number of traces in each group of eventlog
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
Other metrics: 
\code{\link{activity_presence}()},
\code{\link{end_activities}()},
\code{\link{idle_time}()},
\code{\link{number_of_repetitions}()},
\code{\link{processing_time}()},
\code{\link{throughput_time}()},
\code{\link{trace_length}()}
}
