% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{plot}
\alias{plot}
\alias{plot.activity_frequency}
\alias{plot.activity_presence}
\alias{plot.end_activities}
\alias{plot.idle_time}
\alias{plot.processing_time}
\alias{plot.referral_matrix}
\alias{plot.resource_frequency}
\alias{plot.resource_involvement}
\alias{plot.resource_specialisation}
\alias{plot.start_activities}
\alias{plot.throughput_time}
\alias{plot.trace_coverage}
\alias{plot.trace_length}
\alias{plot.number_of_selfloops}
\alias{plot.number_of_repetitions}
\title{Plot Methods}
\usage{
\method{plot}{activity_frequency}(x, ...)

\method{plot}{activity_presence}(x, ...)

\method{plot}{end_activities}(x, ...)

\method{plot}{idle_time}(x, ...)

\method{plot}{processing_time}(x, ...)

\method{plot}{referral_matrix}(x, ...)

\method{plot}{resource_frequency}(x, ...)

\method{plot}{resource_involvement}(x, ...)

\method{plot}{resource_specialisation}(x, ...)

\method{plot}{start_activities}(x, ...)

\method{plot}{throughput_time}(x, ...)

\method{plot}{trace_coverage}(x, ...)

\method{plot}{trace_length}(x, ...)

\method{plot}{number_of_selfloops}(x, ...)

\method{plot}{number_of_repetitions}(x, ...)
}
\arguments{
\item{x}{Data to plot}

\item{...}{Additional variables}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize metric
}
