% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_time.R
\name{processing_time}
\alias{processing_time}
\title{Metric: Processing time}
\usage{
processing_time(eventlog, level_of_analysis, units = "days")
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of processing times should be performed: log, trace, case or activity.}

\item{units}{The time unit in which the throughput times should be reported.}
}
\description{
Provides summary statistics about the processing time of events on the level of activities, traces, cases or log.
}

