% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_repetitions_referral_matrix.R
\name{redo_repetitions_referral_matrix}
\alias{redo_repetitions_referral_matrix}
\alias{redo_repetitions_referral_matrix.eventlog}
\title{Referral matrix repetitons}
\usage{
redo_repetitions_referral_matrix(eventlog)

\method{redo_repetitions_referral_matrix}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}
}
\description{
Provides a list of initatiors and completers of  redo repetitons
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Compute matrix for eventlog
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{\link{number_of_repetitions}}
}
