% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_repetitions_referral_matrix.R
\name{redo_repetitions_referral_matrix}
\alias{redo_repetitions_referral_matrix}
\alias{redo_repetitions_referral_matrix.eventlog}
\title{Referral matrix repetitons}
\usage{
redo_repetitions_referral_matrix(eventlog)

\method{redo_repetitions_referral_matrix}{eventlog}(eventlog)
}
\arguments{
\item{eventlog}{The dataset to be used. Should be a (grouped) eventlog object.
\code{eventlog}.}
}
\description{
Provides a list of initatiors and completers of  redo repetitons
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Compute matrix for eventlog
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{\link{number_of_repetitions}}
}
