% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redo_selfloops_referral_matrix.R
\name{redo_selfloops_referral_matrix}
\alias{redo_selfloops_referral_matrix}
\title{Referral matric selfloops}
\usage{
redo_selfloops_referral_matrix(eventlog)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}
}
\description{
Provides a list of initatiors and completers of  redo selfloops
}
