% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_frequency.R
\name{resource_frequency}
\alias{resource_frequency}
\title{Metric: Resource frequency}
\usage{
resource_frequency(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of  coverage should be performed: log, case, activity, resource, resource-activity.}
}
\description{
Analyses the frequency of resources at different levels of analysis
}

