% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_specialisation.R
\name{resource_specialisation}
\alias{resource_specialisation}
\title{Metric: Resource Specialisation}
\usage{
resource_specialisation(eventlog, level_of_analysis = c("log", "case",
  "resource", "activity"))
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of  coverage should be performed: log, case, resource.}
}
\description{
Analyses whether resources specialise in specific activities
}
