% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdiff.R
\docType{import}
\name{setdiff}
\alias{setdiff}
\alias{setdiff.eventlog}
\title{Set operations}
\usage{
\method{setdiff}{eventlog}(x, y, ...)
}
\arguments{
\item{x}{Eventlog}

\item{y}{Eventlog}
}
\value{
Eventlog of activity instances in x which do not belong to y
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Setdiff for eventlogs
}}

\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{setdiff}}}
}}

