% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_of_repetitions.R
\name{size_of_repetitions}
\alias{size_of_repetitions}
\alias{size_of_repetitions.eventlog}
\alias{size_of_repetitions.grouped_eventlog}
\title{Metric: Size of repetitions}
\usage{
size_of_repetitions(eventlog, type, level, append, ...)

\method{size_of_repetitions}{eventlog}(eventlog, type = c("repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = FALSE, ...)

\method{size_of_repetitions}{grouped_eventlog}(eventlog, type = c("repeat",
  "redo"), level = c("log", "case", "activity", "resource",
  "resource-activity"), append = FALSE, ...)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{type}{The type of selfloops, either repeat or redo.}

\item{level}{At which level the analysis of selfloops should be performed: log, case, activity, resource or resource-activity.}

\item{append}{Logical indicating whether the result should be appended to the orignal event log}

\item{...}{Deprecated arguments}
}
\description{
Provides summary statistics on the sizes of repetitions at the level of activity types, cases, traces or log. A selfloop of size x refers to the occurrence of x consecutive events
of that activity type.
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Size of repetitions for eventlog

\item \code{grouped_eventlog}: Size of repetitions for grouped event log
}}

