% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_of_selfloops.R
\name{size_of_selfloops}
\alias{size_of_selfloops}
\alias{size_of_selfloops.eventlog}
\alias{size_of_selfloops.grouped_eventlog}
\title{Metric: Size of selfloops}
\usage{
size_of_selfloops(eventlog, type, level, append, ...)

\method{size_of_selfloops}{eventlog}(
  eventlog,
  type = c("all", "repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-activity"),
  append = FALSE,
  append_column = NULL,
  ...
)

\method{size_of_selfloops}{grouped_eventlog}(
  eventlog,
  type = c("repeat", "redo"),
  level = c("log", "case", "activity", "resource", "resource-acitivty"),
  append = FALSE,
  append_column = NULL,
  ...
)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{type}{The type of repetitions, either all, repeat or redo.}

\item{level}{Level of granularity for the analysis: log,  case, activity, resource or resource-activity.
For more information, see \code{vignette("metrics", "edeaR")}}

\item{append}{\code{\link{logical}} (default \code{TRUE}): Indicating whether to append results to original log. Ignored when level is \code{log} or \code{trace}.}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated arguments}

\item{append_column}{Which of the output columns to append to log, if \code{append = TRUE}. Default column depends on chosen level.}
}
\description{
Provides summary statistics on the sizes of selfloops
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Size of selfloops for eventlog

\item \code{grouped_eventlog}: Size of selfloops for grouped eventlog
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
\code{\link{number_of_selfloops}}
}
