% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/size_of_selfloops.R
\name{size_of_selfloops}
\alias{size_of_selfloops}
\title{Metric: Size of selfloops}
\usage{
size_of_selfloops(eventlog, level_of_analysis, include_non_selfloops = FALSE)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of selfloops should be performed: log, case, trace or activity.}

\item{include_non_selfloops}{\code{Logical}. When true, also singular events, i.e. selfloops of size 1, are considered.}
}
\description{
Provides summary statistics on the sizes of selfloops at the level of activity types, cases, traces or log. A selfloop of size x refers to the occurrence of x consecutive events
of that activity type.
}

