% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_activities.R
\name{start_activities}
\alias{start_activities}
\title{Metric: Start activities}
\usage{
start_activities(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of start activities should be performed: log, case, activity, resource and resource-activity.}
}
\description{
At log level, computes how many activity types occur as the first event in a case, both absolute and relative.
At activity level, shows the activities which occur as first, and how often.
The first event in a case is the one which started the first.
#'
}

