% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/throughput_time.R
\name{throughput_time}
\alias{throughput_time}
\title{Metric: Throughput time of cases}
\usage{
throughput_time(eventlog, units = "days", level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{units}{The time unit in which the throughput times should be reported.}

\item{level_of_analysis}{At which level the analysis of throughput times should be performed: log or trace.}
}
\description{
Provides summary statistics concerning the throughput times of cases.
The throughput time of cases is defined as the time between the start of the first event and the completion of the last event.
Can be performed at the level of the log as well as the level of traces.
}

