% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_coverage.R
\name{trace_coverage}
\alias{trace_coverage}
\alias{trace_coverage.eventlog}
\alias{trace_coverage.grouped_eventlog}
\title{Metric: Trace coverage}
\usage{
trace_coverage(
  eventlog,
  level = c("log", "trace", "case"),
  append = F,
  append_column = NULL,
  sort = TRUE,
  ...
)

\method{trace_coverage}{eventlog}(
  eventlog,
  level = c("log", "trace", "case"),
  append = F,
  append_column = NULL,
  sort = TRUE,
  ...
)

\method{trace_coverage}{grouped_eventlog}(
  eventlog,
  level = c("log", "trace", "case"),
  append = F,
  append_column = NULL,
  sort = TRUE,
  ...
)
}
\arguments{
\item{eventlog}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}; please use \code{log} instead.}

\item{level}{\code{\link{character}} (default \code{"log"}): Level of granularity for the analysis: \code{"log"} (default),
\code{"trace"}, or \code{"case"}. For more information, see \code{vignette("metrics", "edeaR")} and 'Details' below.}

\item{append}{\code{\link{logical}} (default \code{FALSE}) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}: Indicating whether to append
results to original log. Ignored when level is \code{"log"} or \code{"trace"}.}

\item{append_column}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Which of the output columns to append to log, if \code{append = TRUE}. Default column depends on chosen level.}

\item{sort}{\code{\link{logical}} (default \code{TRUE}): Sort output on count. Only for levels with frequency count output.}

\item{...}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Deprecated arguments}
}
\description{
Analyses the structuredness of an event log by use of trace frequencies. Applicable at logn case and trace level
}
\details{
\itemize{
\item Trace: The absolute and relative frequency of each trace is returned

\item Case: for each case, the coverage of the corresponding trace is returned

\item Log: Summary statistics of the coverage of traces is returned.
}
}
\section{Methods (by class)}{
\itemize{
\item \code{eventlog}: Trace coverage metric for eventlog

\item \code{grouped_eventlog}: Trace coverage metric for grouped eventlog
}}

\references{
Swennen, M. (2018). Using Event Log Knowledge to Support Operational Exellence Techniques (Doctoral dissertation). Hasselt University.
}
\seealso{
Other metrics: 
\code{\link{activity_presence}()},
\code{\link{end_activities}()},
\code{\link{idle_time}()},
\code{\link{number_of_repetitions}()},
\code{\link{number_of_selfloops}()},
\code{\link{processing_time}()},
\code{\link{throughput_time}()},
\code{\link{trace_length}()}
}
