% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_length.R
\name{trace_length}
\alias{trace_length}
\title{Metric: Trace length}
\usage{
trace_length(eventlog, level_of_analysis)
}
\arguments{
\item{eventlog}{The event log to be used. An object of class
\code{eventlog}.}

\item{level_of_analysis}{At which level the analysis of  trace_length should be performed: log, case or trace.}
}
\description{
Computes the length of each trace, in terms of the number of events, at the level of the eventlog or the level of a trace.
The relative numbers at trace level measure trace length compared to the average trace length of the top 80% cases, approximately.
}

